+++ stories/symbolsComponentsMapping/storybookMapping.js
@@ -7,6 +7,7 @@ import {
   notificationsComponentsNames,
   tooltipPopoverComponentsNames,
   modalsComponentsNames,
+  sharedComponents,
 } from './components';
 
 import {
@@ -31,6 +32,7 @@ const componentActualUrl = {
   [layoutComponentsNames.PageHeader]: linkTo('Components', 'PageHeader'),
   [layoutComponentsNames.Layout]: linkTo('Components/Layout', 'Usage'),
   [tooltipPopoverComponentsNames.PopoverMenu]: linkTo('BETA', 'PopoverMenu'),
+  [sharedComponents.Icon]: linkTo('1. Foundation', '1.4 Icons'),
 };
 
 const symbolActualUrl = {